requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
pytorch=="Def"
tensorflow=="Def"
Sqlite=="Def"
Datasets=="Def"
Tranformers=="Def"
pandas=="Def"
numpy=="Def"
tzdata==2023.3
langdetect==1.0.9